[FEAT] 기본 기능 구현 완료

+++ src/main/java/com/konkuk/kubit/domain/Transaction.java
@@ -33,7 +33,7 @@ public class Transaction {
 
     private String resultType;
 
-    private double charge;
+    private int charge;
 
     private double requestPrice;
